# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/matches": {
      "get": {
        "operationId": "Matches_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "scheduled",
                "live",
                "completed",
                "cancelled",
                "postponed"
              ]
            }
          },
          {
            "name": "tournament_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "team_id",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featured",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "scheduled_at",
                "viewer_count"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
...
    "/api/v1/matches/{id}": {
      "get": {
        "operationId": "Matches_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "matches"
        ]
      }
    },
...
    "/api/v1/admin/matches": {
      "post": {
        "operationId": "Admin_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/matches/{id}": {
      "patch": {
        "operationId": "Admin_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "Admin_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/matches/{id}/events": {
      "put": {
        "operationId": "Admin_upsertEvents",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertEventsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/matches/{id}/equipment": {
      "put": {
        "operationId": "Admin_upsertEquipment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertEquipmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/matches/{id}/broadcasts": {
      "put": {
        "operationId": "Admin_upsertBroadcasts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertBroadcastsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/matches/{id}/statistics": {
      "put": {
        "operationId": "Admin_upsertStatistics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertStatisticsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/matches/{id}/economy": {
      "put": {
        "operationId": "Admin_upsertEconomy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertEconomyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/teams": {
      "post": {
        "operationId": "AdminTeams_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/teams/{id}": {
      "patch": {
        "operationId": "AdminTeams_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminTeams_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/matches/{id}/comments": {
      "get": {
        "operationId": "MatchComments_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 30,
              "type": "number"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "matches"
        ]
      },
      "post": {
        "operationId": "MatchComments_create",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "matches"
        ]
      }
    },
    "/api/v1/me/comments/{id}": {
      "delete": {
        "operationId": "MyComments_removeOwn",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/admin/comments/{id}": {
      "delete": {
        "operationId": "AdminComments_removeAny",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "THUMBZ API",
    "description": "Mobile Legends esports catalog, live match data, comments, ticketing and payments. Human-written rules and rationale live in docs/API-CONTRACT.md.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase access token"
      }
    },
    "schemas": {
...
      "CreateMatchDto": {
        "type": "object",
        "properties": {
          "tournament_id": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "stage": {
            "type": "string",
            "nullable": true,
            "enum": [
              "group_stage",
              "regular_season",
              "playoffs",
              "semifinal",
              "third_place",
              "grand_final"
            ]
          },
          "round": {
            "type": "number",
            "nullable": true
          },
          "group_name": {
            "type": "string",
            "nullable": true
          },
          "best_of": {
            "type": "number",
            "minimum": 1
          },
          "game_number": {
            "type": "number",
            "nullable": true
          },
          "team_a_id": {
            "type": "string",
            "format": "uuid"
          },
          "team_b_id": {
            "type": "string",
            "format": "uuid"
          },
          "score_a": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "score_b": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "winner_team_id": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "live",
              "completed",
              "cancelled",
              "postponed"
            ]
          },
          "scheduled_at": {
            "type": "string"
          },
          "started_at": {
            "type": "string",
            "nullable": true
          },
          "ended_at": {
            "type": "string",
            "nullable": true
          },
          "stream_url": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "thumbnail_url": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "viewer_count": {
            "type": "number",
            "minimum": 0
          },
          "featured": {
            "type": "boolean"
          }
        },
        "required": [
          "team_a_id",
          "team_b_id",
          "scheduled_at"
        ]
      },
      "UpdateMatchDto": {
        "type": "object",
        "properties": {
          "stream_delay_seconds": {
            "type": "number",
            "minimum": 0,
            "maximum": 600
          },
          "tournament_id": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "stage": {
            "type": "string",
            "nullable": true,
            "enum": [
              "group_stage",
              "regular_season",
              "playoffs",
              "semifinal",
              "third_place",
              "grand_final"
            ]
          },
          "round": {
            "type": "number",
            "nullable": true
          },
          "group_name": {
            "type": "string",
            "nullable": true
          },
          "best_of": {
            "type": "number",
            "minimum": 1
          },
          "game_number": {
            "type": "number",
            "nullable": true
          },
          "team_a_id": {
            "type": "string",
            "format": "uuid"
          },
          "team_b_id": {
            "type": "string",
            "format": "uuid"
          },
          "score_a": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "score_b": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "winner_team_id": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "live",
              "completed",
              "cancelled",
              "postponed"
            ]
          },
          "scheduled_at": {
            "type": "string"
          },
          "started_at": {
            "type": "string",
            "nullable": true
          },
          "ended_at": {
            "type": "string",
            "nullable": true
          },
          "stream_url": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "thumbnail_url": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "viewer_count": {
            "type": "number",
            "minimum": 0
          },
          "featured": {
            "type": "boolean"
          }
        }
      },
      "MatchEventDto": {
        "type": "object",
        "properties": {
          "team_id": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "player_id": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "event_type": {
            "type": "string",
            "maxLength": 50
          },
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "occurred_at": {
            "type": "string"
          }
        },
        "required": [
          "event_type",
          "title"
        ]
      },
      "UpsertEventsDto": {
        "type": "object",
        "properties": {
          "events": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchEventDto"
            }
          }
        },
        "required": [
          "events"
        ]
      },
      "ItemPurchaseDto": {
        "type": "object",
        "properties": {
          "player_id": {
            "type": "string",
            "format": "uuid"
          },
          "team_id": {
            "type": "string",
            "format": "uuid"
          },
          "item_id": {
            "type": "string"
          },
          "item_name": {
            "type": "string"
          },
          "phase": {
            "type": "string",
            "enum": [
              "phase2",
              "phase3"
            ]
          },
          "slot": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "purchased_at": {
            "type": "string"
          }
        },
        "required": [
          "player_id",
          "team_id",
          "item_id",
          "item_name",
          "phase"
        ]
      },
      "UpsertEquipmentDto": {
        "type": "object",
        "properties": {
          "purchases": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemPurchaseDto"
            }
          }
        },
        "required": [
          "purchases"
        ]
      },
...
      "BroadcastDto": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "enum": [
              "id",
              "en",
              "ms",
              "tl"
            ]
          },
          "stream_url": {
            "type": "string",
            "format": "uri"
          },
          "viewer_count": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "language",
          "stream_url"
        ]
      },
      "UpsertBroadcastsDto": {
        "type": "object",
        "properties": {
          "broadcasts": {
            "maxItems": 8,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BroadcastDto"
            }
          }
        },
        "required": [
          "broadcasts"
        ]
      },
...
      "TeamStatsDto": {
        "type": "object",
        "properties": {
          "team_id": {
            "type": "string",
            "format": "uuid"
          },
          "kills": {
            "type": "number",
            "minimum": 0
          },
          "deaths": {
            "type": "number",
            "minimum": 0
          },
          "assists": {
            "type": "number",
            "minimum": 0
          },
          "gold": {
            "type": "number",
            "minimum": 0
          },
          "towers_destroyed": {
            "type": "number",
            "minimum": 0
          },
          "game_duration_seconds": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "team_id"
        ]
      },
      "PlayerStatsDto": {
        "type": "object",
        "properties": {
          "player_id": {
            "type": "string",
            "format": "uuid"
          },
          "team_id": {
            "type": "string",
            "format": "uuid"
          },
          "kills": {
            "type": "number",
            "minimum": 0
          },
          "deaths": {
            "type": "number",
            "minimum": 0
          },
          "assists": {
            "type": "number",
            "minimum": 0
          },
          "gold": {
            "type": "number",
            "minimum": 0
          },
          "damage": {
            "type": "number",
            "minimum": 0
          },
          "damage_taken": {
            "type": "number",
            "minimum": 0
          },
          "level": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "hero_picked": {
            "type": "string",
            "nullable": true
          },
          "mvp": {
            "type": "boolean"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "player_id",
          "team_id"
        ]
      },
      "UpsertStatisticsDto": {
        "type": "object",
        "properties": {
          "teams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamStatsDto"
            }
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayerStatsDto"
            }
          }
        }
      },
...
      "UpsertEconomyDto": {
        "type": "object",
        "properties": {
          "snapshots": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoldSnapshotDto"
            }
          }
        },
        "required": [
          "snapshots"
        ]
      },
...
      "CreateTeamDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 100,
            "pattern": "^[a-z0-9-]+$"
          },
          "name": {
            "type": "string"
          },
          "short_name": {
            "type": "string",
            "nullable": true,
            "maxLength": 40
          },
          "region": {
            "type": "string"
          },
          "logo_url": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "color_primary": {
            "type": "string",
            "nullable": true,
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "color_secondary": {
            "type": "string",
            "nullable": true,
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "founded_year": {
            "type": "number",
            "nullable": true,
            "minimum": 1900,
            "maximum": 2100
          },
          "is_active": {
            "type": "boolean"
          }
        },
        "required": [
          "slug",
          "name",
          "region"
        ]
      },
      "UpdateTeamDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 100,
            "pattern": "^[a-z0-9-]+$"
          },
          "name": {
            "type": "string"
          },
          "short_name": {
            "type": "string",
            "nullable": true,
            "maxLength": 40
          },
          "region": {
            "type": "string"
          },
          "logo_url": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "color_primary": {
            "type": "string",
            "nullable": true,
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "color_secondary": {
            "type": "string",
            "nullable": true,
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "founded_year": {
            "type": "number",
            "nullable": true,
            "minimum": 1900,
            "maximum": 2100
          },
          "is_active": {
            "type": "boolean"
          }
        }
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 280
          }
        },
        "required": [
          "body"
        ]
      }
    }
  }
}